Don't allow clients to pass a RNG
The default RNG exposed by the math/rand package uses a locked source,
but in general RNGs generated by rand.Source are not safe for use by
multiple goroutines, and the default one exposed by goforit was no
exception.

Instead of exposing this foot-gun, and since nobody (not even the tests)
requires configuring this parameter, just remove it entirely.